third_party/blink/web_tests/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-align-content-005-expected.txt

This is a testharness.js-based test.
PASS .container > div 1
FAIL .container > div 2 assert_equals: 
<div class="container" style="align-content: baseline"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 6
FAIL .container > div 3 assert_equals: 
<div class="container" style="align-content: last baseline"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 6
FAIL .container > div 4 assert_equals: 
<div class="container" style="align-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 6
PASS .container > div 5
PASS .container > div 6
PASS .container > div 7
PASS .container > div 8
FAIL .container > div 9 assert_equals: 
<div class="container" style="align-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 6
FAIL .container > div 10 assert_equals: 
<div class="container" style="align-content: end"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 6
FAIL .container > div 11 assert_equals: 
<div class="container" style="align-content: flex-start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 6
FAIL .container > div 12 assert_equals: 
<div class="container" style="align-content: flex-end"><div data-offset-x="10" data-offset-y="1"></div></div>
offsetLeft expected 10 but got 6
PASS .container > div 13
FAIL .container > div 14 assert_equals: 
<div class="container" style="align-content: baseline"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 0
FAIL .container > div 15 assert_equals: 
<div class="container" style="align-content: last baseline"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 0
FAIL .container > div 16 assert_equals: 
<div class="container" style="align-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 0
PASS .container > div 17
PASS .container > div 18
PASS .container > div 19
PASS .container > div 20
FAIL .container > div 21 assert_equals: 
<div class="container" style="align-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 0
FAIL .container > div 22 assert_equals: 
<div class="container" style="align-content: end"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 0
FAIL .container > div 23 assert_equals: 
<div class="container" style="align-content: flex-start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetLeft expected 2 but got 0
FAIL .container > div 24 assert_equals: 
<div class="container" style="align-content: flex-end"><div data-offset-x="-2" data-offset-y="1"></div></div>
offsetLeft expected -2 but got 0
Harness: the test ran to completion.
